drag and drop on edit property fragment working fine, just needs refactoring

+++ app/src/main/java/almeida/francisco/forestboundaries/LocatePropertyFragment.java
@@ -162,7 +162,7 @@ public class LocatePropertyFragment
             public void onClick(View view) {
                 for (int i = 0; i < points.size(); i++) {
                     markerService.createMarker(new MyMarker()
-                            .setIndex((double) i)
+                            .setIndex(i)
                             .setMarkedLatitude(points.get(i).latitude)
                             .setMarkedLongitude(points.get(i).longitude)
                             .setProperty(property));
